Allow only 1 focused list
When using the arrow keys to navigate between
transaction list and person list, only one of the list
items can be selected.

For example, using the right arrow key -> to navigate to the
transaction list from the person list will cause the previously
selected person card to be unselected.

Changes made:

Added an unFocus method to UiPartFocusable